api/requirements.txt
Django==3.1.1
djangorestframework
uwsgi
psycopg2
celery